* removing extCookie ajax call from js * removing from yaml * the extCookie dynamic session moved from gigya controller to gigyaRaasHelper and been called from eventSubscriber
